--- website/src/components/issues/ColumnDefs.js
@@ -1,97 +1,0 @@
-const AllColumns = {
-  Repo: {
-    title: "Repo",
-    display: true,
-    filter: null,
-    selection: null,
-  },
-  Issue: {
-    title: "Issue",
-    display: true,
-    filter: null,
-    selection: null,
-  },
-  Title: {
-    title: "Title",
-    display: true,
-    filter: null,
-    selection: null,
-  },
-  Created: {
-    title: "Created",
-    display: true,
-    filter: null,
-    selection: null,
-  },
-  Updated: {
-    title: "Updated",
-    display: true,
-    filter: null,
-    selection: null,
-  },
-  ClosedAt: {
-    title: "Closed",
-    display: true,
-    filter: null,
-    selection: null,
-  },
-  ClosedBy: {
-    title: "Closed By",
-    display: true,
-    filter: null,
-    selection: null,
-  },
-  State: {
-    title: "State",
-    display: true,
-    filter: null,
-    selection: null,
-  },
-  LinkedPR: {
-    title: "Linked PR",
-    display: true,
-    filter: null,
-    selection: null,
-  },
-  Assignee: {
-    title: "Assignee",
-    display: true,
-    filter: null,
-    selection: null,
-  },
-  Affects: {
-    title: "Affects",
-    display: true,
-    filter: null,
-    selection: null,
-    columns: [
-      {
-        title: "Version",
-        display: true,
-      },
-      {
-        title: "Affects",
-        display: true,
-      },
-      {
-        title: "PR",
-        display: true,
-      },
-      {
-        title: "State",
-        display: true,
-      },
-      {
-        title: "Release",
-        display: true,
-      },
-    ],
-  },
-  Severity: {
-    title: "Severity",
-    display: true,
-    filter: null,
-    selection: null,
-  },
-};
-export default AllColumns;
