+++ website/src/pages/open.js
@@ -1,32 +1,56 @@
-import * as React from "react";
-import Container from "@mui/material/Container";
-import Layout from "../layout/Layout";
-
-import Tab from "@mui/material/Tab";
-import {
-  Accordion,
-  AccordionDetails,
-  AccordionSummary,
-  Link,
-} from "@mui/material";
-import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
+import { AccordionDetails, AccordionSummary, Container } from "@mui/material";
+import React from "react";
+import { Layout } from "../layout/Layout";
+import Accordion from "@mui/material/Accordion";
+import ExpandMore from "@mui/icons-material/ExpandMore";
 import Tabs from "@mui/material/Tabs";
-import Box from "@mui/material/Box";
-
-import { useQuery } from "react-query";
-import { url } from "../utils";
+import Tab from "@mui/material/Tab";
 import { IssueGrid } from "../components/issues/IssueGrid";
 import Columns from "../components/issues/GridColumns";
+import { currentVersions, url } from "../utils";
+import { useQuery } from "react-query";
 import {
-  affectState,
-  repo,
   severity,
   state,
-  hasPR,
-  noPR,
+  type,
+  openIn24h,
+  openSince,
+  NOT,
+  AND,
 } from "../components/issues/filter/index";
 
-function OpenedToday() {
+function RecentOpen() {
+  const filters = [
+    type("bug"),
+    AND([NOT(severity("moderate")), NOT(severity("minor"))]),
+    state("open"),
+  ];
+  const pickColumns = [];
+  for (const version of currentVersions) {
+    pickColumns.push(Columns.getAffectionOnVersion(version));
+  }
+
+  const [tab, setTab] = React.useState(0);
+  const tabs = ["Created in 24h", "Created in 7d", "Created in 30d", "ALL"];
+  switch (tab) {
+    case 0:
+      filters.push(openIn24h());
+      break;
+    case 1:
+      filters.push(openSince(new Date().getTime() - 60 * 60 * 1000 * 24 * 7));
+      break;
+    case 2:
+      filters.push(openSince(new Date().getTime() - 60 * 60 * 1000 * 24 * 30));
+      break;
+    case 3:
+      break;
+    default:
+      break;
+  }
+
+  const handleChange = (event, newValue) => {
+    setTab(newValue);
+  };
   const { isLoading, error, data } = useQuery("issue", () => {
     return fetch(url("issue"))
       .then((res) => {
@@ -37,100 +61,41 @@ function OpenedToday() {
         console.log(e);
       });
   });
-  console.log(isLoading, error, data);
-  if (isLoading) {
-    return (
-      <div>
-        <p>Loading...</p>
-      </div>
-    );
-  }
-  if (error) {
-    return (
-      <div>
-        <p>Error: {error.message}</p>
-      </div>
-    );
-  }
-  console.log("fetched data", data);
-  return (
-    <IssueGrid
-      data={data.data}
-      columns={[
-        Columns.repo,
-        Columns.number,
-        Columns.title,
-        Columns.state,
-        Columns.pr,
-        Columns.type,
-        Columns.severity,
-        Columns.labels,
-        Columns.getAffectionOnVersion("5.4"),
-        Columns.getPROnVersion("5.4"),
-        Columns.getPickOnVersion("5.4"),
-        Columns.getAffectionOnVersion("5.3"),
-        Columns.getPROnVersion("5.3"),
-        Columns.getPickOnVersion("5.3"),
-        Columns.getAffectionOnVersion("5.2"),
-        Columns.getPROnVersion("5.2"),
-        Columns.getPickOnVersion("5.2"),
-        Columns.getAffectionOnVersion("5.1"),
-        Columns.getPROnVersion("5.1"),
-        Columns.getPickOnVersion("5.1"),
-        Columns.getAffectionOnVersion("5.0"),
-        Columns.getPROnVersion("5.0"),
-        Columns.getPickOnVersion("5.0"),
-        Columns.getAffectionOnVersion("4.0"),
-        Columns.getPROnVersion("4.0"),
-        Columns.getPickOnVersion("4.0"),
-      ]}
-      filters={[
-        repo("tidb"),
-        state("closed"),
-        // severity("critical"),
-        affectState("5.3", "yes"),
-        hasPR("master"),
-        hasPR("release-5.3"),
-      ]}
-    ></IssueGrid>
-  );
-}
-
-const RecentOpen = () => {
-  const [value, setValue] = React.useState(0);
-
-  const handleChange = (event, newValue) => {
-    setValue(newValue);
-  };
-
   return (
     <Layout>
       <Container maxWidth="xxl" sx={{ mt: 4, mb: 4 }}>
         <Accordion defaultExpanded={true}>
-          <AccordionSummary expandIcon={<ExpandMoreIcon />}>
+          <AccordionSummary expandIcon={<ExpandMore />}>
             Recent Open
           </AccordionSummary>
           <AccordionDetails>
-            <Box sx={{ width: "100%" }}>
-              <Box sx={{ borderBottom: 1, borderColor: "divider" }}>
-                <Tabs
-                  value={value}
-                  onChange={handleChange}
-                  aria-label="basic tabs example"
-                >
-                  <Tab label="Opened Today" />
-                  <Tab label="Opened This Week" />
-                  <Tab label="Opened This Month" />
-                  <Tab label="All Open Issues" />
-                </Tabs>
-              </Box>
-              <OpenedToday />
-            </Box>
+            <Tabs value={tab} onChange={handleChange}>
+              {tabs.map((v) => (
+                <Tab label={v}></Tab>
+              ))}
+            </Tabs>
+            {isLoading && <p>Loading...</p>}
+            {error && <p>Error: {error.message}</p>}
+            {data && (
+              <IssueGrid
+                columns={[
+                  Columns.repo,
+                  Columns.number,
+                  Columns.title,
+                  Columns.type,
+                  Columns.severity,
+                  Columns.state,
+                  ...pickColumns,
+                ]}
+                data={data.data}
+                filters={filters}
+              ></IssueGrid>
+            )}
           </AccordionDetails>
         </Accordion>
       </Container>
     </Layout>
   );
-};
+}
 
 export default RecentOpen;
